function/requirements.txt
azure-functions
psycopg2-binary
isort==5.7.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
pylint==2.6.0
six==1.15.0
toml==0.10.2
typed-ast==1.4.2
wrapt==1.12.1
datetime
sendgrid
python-http-client==3.3.1
astroid==2.4.2
starkbank-ecdsa==1.1.0
colorama==0.4.4